test(AvatarGroupOverflow): add component-scoped audit coverage
